move actors into separate folder
