+++ src/compiler/file_uploader/index.ts
@@ -2,7 +2,7 @@ import { getCanisterId } from '../../../dfx';
 import { expandPaths } from './expand_paths';
 import { uploadFile } from './upload_file';
 import { onBeforeExit } from './on_before_exit';
-import { createActor } from './file_chunk_actor';
+import { createActor } from './actors/file_chunk_actor';
 
 export type Src = string;
 export type Dest = string;
